Bump OpenTelemetry deps to 1.22.0
OpenTelemetry 1.22.0 introduces a breaking change for manually
instrumented applications that use autoconfiguration. `clj-otel`
relies on the global OpenTelemetry instance to be set on first `get` in
this particular use case. This behaviour is now opt-in for Java
OpenTelemetry 1.22.0 and requires a system property (or environment
variable) to be set.

Documentation for migration has been added to CHANGELOG.adoc