Allow `observe` fn to take several measurements
The `instrument` function supports instruments that take asynchronous
measurements by using the provided 0-arity function `observe`.

Previously, `observe` would be expected to return a single map of
a measurement to record. The change in this commit allows `observe` to
return a sequence of maps for recording several measurements. This may
be used to simultaneously record measurements with different attributes.

Fixes #5