Add bound context as default context parameter
Bound context is a new opt-in feature which leverages a Clojure dynamic
var. It overrides current context as the default `context` or `parent`
parameter value in `clj-otel` functions. This is intended to simplify
async code by eliminating explicit context passing.

Existing code that uses the current context or explicit context
parameters is unaffected by the introduction of the bound context.