Remove use of EXCEPTION_ESCAPED semantic attribute
Future versions of OpenTelemetry Java do not have the semantic
attribute `ExceptionAttributes/EXCEPTION_ESCAPED`. To retain
compatibility, this commit removes references to this attribute.

OpenTelemetry no longer recommends to record exceptions that are handled
and do not escape the scope of a span. Functions `wrap-exception-event`
and `exception-event-interceptor` are now deprecated and will be stubbed
out and later removed in future `clj-otel` releases.