Generalise context propagation functions
Add options to `->headers` and `headers->merged-context` for use
in applications other than Ring HTTP servers.

Options `:getter` and `:setter` have been added to generalise the use
of these functions. The defaults are suitable for use in a Ring HTTP
server application. In particular, the default `:getter` lower-cases
the key before lookup in `headers`. The Ring spec states that keys
in a request `headers` map are in lower case.

Fixes #26