Bump Otel 1.50.0 and instrumentation 2.16.0-alpha
Bump
- OpenTelemetry deps to 1.50.0
- OpenTelemetry instrumentation deps to 2.16.0-alpha

Note: Otel 1.50.0 is compatible with agent 2.16.0 and later. Earlier
versions of the agent may give exceptions complaining about missing
class `ExtendedAttributes` when using log statements that are captured
by OpenTelemetry. This is due to a breaking change
introduced by
https://github.com/open-telemetry/opentelemetry-java/pull/7123
If it is not possible to use this agent version, add dependency
`io.opentelemetry/opentelemetry-api-incubator`.