Populate `code.function.name` attribute for spans
This adds the name of the function containing the start of each span
to the span data. It provides a default for the `:fn` entry of the
`:source` map in `span-opts`. The function name is computed at
runtime when spans are started, as this information is not available
at macro expansion time.

This is only added when the StackWalker API is available, which was
first introduced in JDK 9. Techniques predating StackWalker involve
access of the entire stack trace, which is slow for deep stacks e.g.
`(doto (Throwable.) .fillInStackTrace)`.