Ensure HTTP route data is provided to agent
When the OpenTelemetry instrumentation agent is used, HTTP route
data for server spans need to be provided to the agent to add as
an attribute to server span metrics e.g. `http.route` attribute on
request duration metrics.

OpenTelemetry Java publishes a module
`io.opentelemetry.instrumentation/opentelemetry-instrumentation-api`.
This includes `HttpServerRoute` which enables applications to update the
route data. When the agent is not present, this update method does
nothing.