Add `:set-bound-context?` option to interceptor
Previously, `server-span-interceptors` would always set the bound
context. This would not allow correct usage of current context in
subsequent processing, as the bound context would override it.

A new option `:set-bound-context?` has been added to
`server-span-interceptors`, with default false. It works for new and
existing server spans. To use the bound context in subsequent
processing, this option must be set to true.