Fix `MeterProvider` for OTLP exporters
This fixes an issue where the global OpenTelemetry is initialized
during programmatic building of an OTLP exporter.

This arises from a behavioral change in recent OpenTelemetry Java
releases, where a Supplier<MeterProvider> is given as an input to
OTLP exporters, to monitor exporter performance. The default for this
supplier is taken from GlobalOpenTelemetry. The action of using this
supplier has the side effect of initializing the global OpenTelemetry
instance.

This commit avoids the unwanted initialization by providing an
alternative default supplier, where the default OpenTelemetry instance
is used instead.