Add `clj-otel-adapter-log4j` module
This commit adds a new module `clj-otel-adapter-log4j`, which provides:
- A Log4j appender that emits log records
- A Log4j context data provider that is clj-otel bound context aware

This is an alternative to OpenTelemetry artifacts
`io.opentelemetry.instrumentation/opentelemetry-log4j-appender-2.17`
and
`io.opentelemetry.instrumentation/opentelemetry-log4j-context-data-2.17`

`clj-otel-adapter-log4j` has the same configuration options as the
OpenTelemetry counterparts. However, the name of the appender and
context data keys are different. This is to avoid conflicts, as it is
not possible to remove or mute the OpenTelemetry appender and context
data provider when using the OpenTelemetry instrumentation agent.

`clj-otel-adapter-log4j` provides the same functionality as
OpenTelemetry artifacts, but is bound context aware. Using this
appender ensures the correct context is attached to each log event by
default, when the application uses current or bound context.