Add `clj-otel-adapter-logback` module
This commit adds a new module `clj-otel-adapter-logback`, which
provides:
- A Logback appender that emits log records
- A Logback appender that adds Otel context data to Logback MDC. This
appender is designed to be used to wrap other appenders, such as the log
record emitter appender described above.

This is an alternative to OpenTelemetry artifacts
`io.opentelemetry.instrumentation/opentelemetry-logback-appender-1.0`
and
`io.opentelemetry.instrumentation/opentelemetry-logback-mdc-1.0`

`clj-otel-adapter-logback` has the same configuration options as the
OpenTelemetry counterparts. To use with the OpenTelemetry agent,
the agent's Logback instrumentation must be disabled to prevent
conflicts. Use these JVM flags:
-Dotel.instrumentation.logback-appender.enabled=false
-Dotel.instrumentation.logback-mdc.enabled=false

`clj-otel-adapter-log4j` provides the same functionality as
OpenTelemetry artifacts, but is bound context aware. Using this
appender ensures the correct context is attached to each log event by
default, when the application uses current or bound context.