Add other forms for `:logger` in `log-record/emit`
The `:logger` in `log-record/emit` option map may now be specified as
a map or name string. This change enables evaluations of `emit` to be
delayed by storing the entire option map for later. By delaying
retrieval of the logger until `emit` evaluation, this prevents issues
with early retrieval of the logger before the desired OpenTelemetry
instance has been constructed.